+++ lib/vfs/mc-vfs/ftpfs.c
@@ -1660,8 +1660,7 @@ ftpfs_file_store (struct vfs_class *me, struct vfs_s_fh *fh, char *name, char *l
             w_buf += n_written;
             n_read -= n_written;
         }
-        print_vfs_message (_("ftpfs: storing file %lu (%lu)"),
-                           (unsigned long) n_stored, (unsigned long) s.st_size);
+        print_vfs_message (_("ftpfs: storing file %ju (%ju)"), n_stored, s.st_size);
     }
     tty_disable_interrupt_key ();
     close (sock);
